chore: update haml dependency version
This updates to the haml dependency relative to the current version.
haml 3 has a cross scripting security issue with ruby. To avoid glitter
installing an insecure of version haml where this is the only gem with a
haml dependency, the version is set to ~>5. This also removes the
deprecated `rubyforge_project` property in the gemspec.